crud for user cart

+++ models/cart.go
@@ -3,6 +3,6 @@ package models
 // Cart collection
 type Cart struct {
 	Product     string `bson:"product" json:"product"`
-	Qty         int16  `bson:"qty" json:"qty"`
+	Qty         int    `bson:"qty" json:"qty"`
 	Description string `bson:"description" json:"description"`
 }
